Health and fixed errors
Also change a few things. I added a HP=100 at the sttart function of the creeps because vitalityAttributes doesnt show up on the inspector even though it's public.
Contact me whenever you want to find a fix for it.

+++ 2DPlatformerController/Assets/Attributes/DamagableAttributes.cs
@@ -6,7 +6,7 @@ using System;
 public class DamagableAttributes
 {
     public int AttackDamage;
-    public int AttackDelaySeconds;
+    public float AttackDelaySeconds;
     public DateTime LastAttackTime;
     public decimal Radius;
     public int Cleave;//1 target only
